Making pagination function in user-controller
First, the function call getUsers() to check get the total of users.
After getting the paramter from body and query and setting the default value. The function calls
getUsersPagination in users-service. It will return an object containing page info and data.
If the lenght of the array equals 0, then there is no matching user, return error 404.
Else return the users data and info that match the algorithm